move: types/の中をconfigデータとmodelsに分けた。modelsの中もprefecturesとpopulationCompositionに分けた。またそれに合わせた他のファイルのパス修正
